+++ example/vit_cache_example.dart
@@ -1,19 +1,15 @@
 import 'package:vit_cache/vit_cache.dart';
 
-class UserInfoCache extends SingularCache<Map<String, dynamic>> {
-  @override
-  Future<Map<String, dynamic>> fetch() async {
-    // Simulate a network call to fetch user info
-    await Future.delayed(Duration(seconds: 2));
-    return {'name': 'Dave', 'auth_token': 'xxxxx'};
-  }
-
-  @override
-  Duration get ttl => Duration(seconds: 10);
-}
-
 void main() async {
-  var userInfoCache = UserInfoCache();
+  // Create a cache with a fetch function
+  var userInfoCache = SingularCache<Map<String, dynamic>>(
+    ttl: Duration(seconds: 10),
+    fetch: () async {
+      // Simulate a network call to fetch user info
+      await Future.delayed(Duration(seconds: 2));
+      return {'name': 'Dave', 'auth_token': 'xxxxx'};
+    },
+  );
 
   // Fetch and cache user info
   var info = await userInfoCache.get();