+++ lib/vit_cache.dart
@@ -3,6 +3,8 @@
 /// More dartdocs go here.
 library;
 
-export 'src/cache/multi_timed_cache_model.dart';
-export 'src/cache/singular_cache_model.dart';
+export 'src/cache/multi_timed_cache.dart';
+export 'src/cache/singular_cache.dart';
 export 'src/cache/timed_cache_model.dart';
+export 'src/data/models/cached_item.dart';
+export 'src/errors/cache_item_missing.dart';